fixed some ui problems with JK

+++ src/main/java/kernbeisser/CustomComponents/ShoppingTable/ShoppingCartController.java
@@ -12,7 +12,7 @@ public class ShoppingCartController implements Controller<ShoppingCartView,Shopp
     private ShoppingCartModel model;
 
 
-    public ShoppingCartController(double userValue, int userSurcharge) {
+    public ShoppingCartController(double userValue, double userSurcharge) {
         model = new ShoppingCartModel(userValue, userSurcharge);
         view = new ShoppingCartView(this);
     }
